+++ pkgs/ffigen/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 16.0.0-wip
 
 - Ensure all protocols referenced in bindings are available at runtime.
+- Use `package:dart_style` directly to format generated Dart code, rather than
+  subprocessing to `dart format`.
 - Use package:objective_c 4.0.0
 - Fix various small bugs todo with config filters:
   - https://github.com/dart-lang/native/issues/1582